53 - separated models from reducers, modal styles updated
