+++ apps/lumina-core/src/components/lumina-modals/modals.module.scss
@@ -5,33 +5,15 @@
 }
 
 .modalEdit {
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-.modalOverlay {
-  background-color: color-mix(in srgb, var(--lum-color-background-base) 80%, transparent);
-  position: fixed;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  padding: 40px;
-  box-sizing: border-box;
-  z-index: 2000;
-
-  * {
-    box-sizing: border-box;
-  }
-}
-
-.modalEdit {
   margin: 0 auto;
   width: 50%;
-  min-width: 950px;
-  max-width: 100%;
-  max-height: 700px;
-  border: 1px solid;
-  background: var(--lum-color-background-base);
-  border-radius: 5px;
   padding: 40px;
+  border: 1px solid transparent;
+  background: var(--lum-color-white);
+  border-radius: 10px;
+  box-shadow: 0px 2px 4px -1px rgba(0, 0, 0, 0.06), 0px 4px 6px -1px rgba(0, 0, 0, 0.10);
+  position: absolute;
+  top: 20%;
+  left: 50%;
+  transform: translateX(-50%);
 }
