+++ apps/lumina-core/src/context/reducers/models/builderDataModels.ts
@@ -1,0 +1,12 @@
+import { IData } from "@/data/data";
+
+export interface IBuilderDataContext {
+  builderData: IData;
+  selectedPage: string;
+  pages: string[];
+}
+
+export interface ISetBuilderDataAction {
+  type: "setBuilderData";
+  data: IBuilderDataContext;
+}
